requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
gunicorn==21.2.0
